+++ src/rend/audio/mod.rs
@@ -1,0 +1,6 @@
+// devela::rend::audio
+//
+//! Audio functionality.
+//
+
+pub(crate) mod all {}