+++ src/rend/draw/mod.rs
@@ -1,0 +1,6 @@
+// devela::rend::draw
+//
+//! Drawing functionality.
+//
+
+pub(crate) mod all {}