+++ src/rend/font/mod.rs
@@ -1,0 +1,6 @@
+// devela::rend::font
+//
+//! Font functionality.
+//
+
+pub(crate) mod all {}