+++ src/rend/layout/mod.rs
@@ -1,0 +1,6 @@
+// devela::rend::layout
+//
+//! Layout functionality.
+//
+
+pub(crate) mod all {}